refactor: Floor enum과 BookLocation 클래스의 중복 구현 제거
- Floor enum 중복 제거 (model.entity.Floor -> model.Floor)
- BookLocation 클래스 중복 제거 (model.BookLocation -> model.entity.BookLocation)
- 관련 import 구문 수정
