--- booker-server/src/main/java/com/bookerapp/core/domain/model/entity/Floor.java
@@ -1,19 +1,0 @@
-package com.bookerapp.core.domain.model.entity;
-
-import lombok.Getter;
-
-@Getter
-public enum Floor {
-    FOURTH(4),
-    TWELFTH(12);
-
-    private final int number;
-
-    Floor(int number) {
-        this.number = number;
-    }
-
-    public int getValue() {
-        return number;
-    }
-}
