+++ booker-server/src/test/java/com/bookerapp/core/domain/model/dto/BookDtoTest.java
@@ -3,7 +3,7 @@ package com.bookerapp.core.domain.model.dto;
 import com.bookerapp.core.domain.model.entity.Book;
 import com.bookerapp.core.domain.model.entity.BookLocation;
 import com.bookerapp.core.domain.model.entity.BookStatus;
-import com.bookerapp.core.domain.model.entity.Floor;
+import com.bookerapp.core.domain.model.Floor;
 import jakarta.validation.ConstraintViolation;
 import jakarta.validation.Validation;
 import jakarta.validation.Validator;
